% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone_reporting_module.R
\name{milestone_reporting_module_ui}
\alias{milestone_reporting_module_ui}
\alias{milestone_reporting_module_server}
\title{milestone_reporting_module_ui and milestone_reporting_module_server}
\usage{
milestone_reporting_module_ui(id, button_text = "Download plot table")

milestone_reporting_module_server(id, data, config, syn, study_id, verbose = T)
}
\arguments{
\item{id}{shiny id}

\item{button_text}{button_text A string}

\item{data}{A named list. The list must contain a list named "tables".}

\item{config}{A named list.}
}
\description{
A shiny Module.
}
\keyword{internal}
